- Add IterateExample to iteration example
- Add ReadFromKafka to Kafka example
- Add WriteIntoKafka to Kafka example
- Add IncrementalLearningSkeleton to ml example
- Add TwitterExample to Twitter example
- Add GroupedProcessingTimeWindowExample to windowing example
- Add SessionWindowing to windowing example
- Add WindowWordCount to windowing example
- Add WordCount to wordcount example

[FLINK-3551] [examples] Remove PojoExample from wordcount examples

+++ flink-java/src/test/java/org/apache/flink/api/common/operators/CollectionExecutionAccumulatorsTest.java
@@ -46,8 +46,8 @@ public class CollectionExecutionAccumulatorsTest {
 			JobExecutionResult result = env.execute();
 			
 			assertTrue(result.getNetRuntime() >= 0);
-			
-			assertEquals(NUM_ELEMENTS, result.getAccumulatorResult(ACCUMULATOR_NAME));
+
+			assertEquals(NUM_ELEMENTS, (int) result.getAccumulatorResult(ACCUMULATOR_NAME));
 		}
 		catch (Exception e) {
 			e.printStackTrace();
